refactor(fs): shared the directory node logic between ramfs and devfs
ramfs and devfs each carried their own copy of the child list, name
lookup, directory listing, and attribute reporting, so every directory
behavior had to be written twice and any new directory attribute could
land in one filesystem and silently miss the other. A common directory
base now owns the child list and the reads over it, while each
filesystem keeps only the mutations it chooses to allow, which leaves
ramfs writable and devfs populated solely by kernel drivers. The
non-recursive teardown that ramfs already used becomes the one
destructor, so nested directories in any filesystem unwind with
bounded stack depth.
